Added in home.js for easier merging
